Safari-609.1.11.1/LayoutTests/svg/dom/SVGSVGElement-getElementById.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests that calling SVGSVGElement.getElementById() without parameter throws an exception.");

var svg = document.createElementNS("http://www.w3.org/2000/svg", "svg");
shouldThrow("svg.getElementById()", "'TypeError: Not enough arguments'");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
