Safari-609.1.11.1/LayoutTests/svg/dom/SVGSVGElement-getElementById-expected.txt

Tests that calling SVGSVGElement.getElementById() without parameter throws an exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svg.getElementById() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
